+++ src/Modules/Region/RegionController.php
@@ -296,17 +296,11 @@ final class RegionController extends FoodsharingController
                 return $this->redirect('/region?sub=forum&bid=' . $region['id']);
             }
             $this->pageHelper->addTitle($thread['title']);
-            $pageData['threadId'] = $threadId;
         } elseif ($request->query->has('newthread')) {
             $this->pageHelper->addTitle($this->translator->trans('forum.new_thread'));
-            $postActiveWithoutModeration = $this->forumPermissions->mayStartUnmoderatedThread($region, $ambassadorForum);
-            $pageData['newThreadForm'] = true;
-            $pageData['postActiveWithoutModeration'] = $postActiveWithoutModeration;
-        } else {
-            $pageData['threads'] = [];
         }
 
-        $params = $this->convertDataToObject($region, $sub, $pageData);
+        $params = $this->convertDataToObject($region, $sub, []);
 
         $this->pageHelper->addContent($this->view->vueComponent('region-page', 'RegionPage', $params));